Reorganization of the structure to fit better with SimPEG.
Move the "Problem" to something called "System", and work with the assumption that it's not really public-facing.
Start to move source/receiver stuff to the Survey and out of Kernel (though these may still need to be aware of each other, and in fact we perhaps want the source-term material to be part of UniformMesh).