Introduce a new, much more flexible matrix ordering construction.
The new matrix ordering construction from MiniZephyr is introduced in the form of the 'mord' configuration parameter. The matrix diagonal offsets are computed using the evaluated versions of the first shift and the second shift from the 'mord' parameter, which is a tuple of strings. So, for example, ('+nx', '+1') would result in shifting by +/– nx on the first dimension of the stencil and +/– 1 on the second dimension. A value of ('-1', '-nz') would shift by –/+ 1 on the first dimension and –/+ nz on the second. This means that we can essentially transpose and flip the stencil arbitrarily, as well as accommodate different array orderings from Fortran and/or C.