Add configurable cPML, new default value, update notebook defaults.
The cPML property has been added as a configurable parameter in the Eurus class. Also, from some initial testing, I'm suggesting a new default value of 1000. I've adjusted some of the parameters in the notebook to show this off a bit better (at higher frequency, and on a non-square domain). Also, plot size has been increased.