Add SCFilter and BaseSCCache classes.
SCFilter takes a collection of AttributeMapper classes, and computes the combined requirements and optional arguments. Then, any systemConfig passed through it will be stripped of irrelevant entries, and an error will be raised if a required entry isn't provided. This is useful, for example, before Pickling and sending over a network.

BaseSCCache is a new base class that caches systemConfig, so that this functionality can be inherited without a DiscretizationWrapper.