Support checking / deleting matrix factors
There is now a property called "factors" on Problem, DiscretizationWrapper and Discretization objects. It returns True if there are factors / pools taking up memory, and False if there are none. Calling "del problem.factors" will do a recursive cleanup.

NB: This is a bit of a metaphor; the LU factors are not the only thing deleted, nor are they the only thing that will return True, per se.